+++ utils/constants.js
@@ -5,3 +5,7 @@ export const ITEM_ID = '4'
 export const ITEM_NAME = 'Ticket Pass-Mueseum'
 
 export const CAT_ATTR_TAG_RELATIONS = [2, 3, 4, 5]
+export const TRIGGER_BLIZZARD_MESSAGE = "Hey, Triggering a Blizzard";
+export const CANCEL_BOOKING_MESSAGE = `Dear Guest,\n\nApologies, but your hotel booking with us has been canceled due to unforeseen circumstances. \nWe understand the inconvenience and are here to assist you with any alternative arrangements needed. \n\nPlease contact us for further assistance.`;
+export const NEW_CATALOG_AVAILABLE = `Dear Guest,\n\n Checkout this new place to visit.`
+
